Feat: 매칭글 호스트 수정 기능 구현 완료 - PATCH 메서드에서 PUT 메서드로 수정함. - 클라이언트가 수정 요청시 모든 필드의 내용을 페이로드에 담아 전송하는 것을 가정하여 업데이트 메서드를 수정함.

+++ src/main/java/com/likelion/hufsting/domain/matching/dto/CreateMatchingPostData.java
@@ -12,7 +12,7 @@ public class CreateMatchingPostData {
     private Gender gender;
     private int desiredNumPeople;
     private String openTalkLink;
-    private List<Integer> participants;
+    private List<Long> participants;
 
     // Controller-Service 간 DTO 변환 함수
     public static CreateMatchingPostData toCreateMatchingPostData(CreateMatchingPostRequest dto){
